refactor functions and constants to back-end/constants and back-end/functions
