+++ back-end/constants/FAQData.js
@@ -1,0 +1,21 @@
+/**
+ * Default placeholder FAQ Data. This will be in the DB in the future.
+ */
+const FAQData = {
+    rows: [
+      {
+        title: "Lorem ipsum dolor sit amet 1?",
+        content: "Lorem ipsum dolor sit amet 1",
+      },
+      {
+        title: "Lorem ipsum dolor sit amet 2?",
+        content: "Lorem ipsum dolor sit amet 2",
+      },
+      {
+        title: "Lorem ipsum dolor sit amet 3?",
+        content: "Lorem ipsum dolor sit amet 3",
+      },
+    ],
+};
+  
+module.exports = FAQData;
