+++ back-end/constants/categories.js
@@ -1,0 +1,21 @@
+/**
+ * Default categories for transactions. Custom categories may be moved to the 
+ * database in the future.
+ */
+const categories = [
+    { name: "Bank Fees", icon: "MdAccountBalance" },
+    { name: "Cash Advance", icon: "MdAccountBalance" },
+    { name: "Community", icon: "MdAccountBalance" },
+    { name: "Food and Drink", icon: "MdOutlineLocalCafe" },
+    { name: "Healthcare", icon: "AiOutlineHome" },
+    { name: "Interest", icon: "MdAccountBalance" },
+    { name: "Payment", icon: "MdAccountBalance" },
+    { name: "Recreation", icon: "MdOutlineSportsHandball" },
+    { name: "Service", icon: "MdOutlineCastForEducation" },
+    { name: "Shops", icon: "MdOutlineLocalGroceryStore" },
+    { name: "Tax", icon: "MdAccountBalance" },
+    { name: "Transfer", icon: "MdAccountBalance" },
+    { name: "Travel", icon: "MdEmojiTransportation" },
+];
+
+module.exports = categories;
