+++ back-end/functions/formatError.js
@@ -1,0 +1,7 @@
+const formatError = (error) => {
+    return {
+      error: { ...error.data, status_code: error.status },
+    };
+};
+
+module.exports = formatError;
