+++ back-end/functions/prettyPrintResponse.js
@@ -1,0 +1,6 @@
+const prettyPrintResponse = (response) => {
+    if (response.data) console.log(response.data);
+    else console.log(JSON.stringify(response));
+};
+
+module.exports = prettyPrintResponse;
